python-backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2024.12.14
click==8.1.8
distro==1.9.0
Flask==3.1.0
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
openai==1.60.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
python-dotenv==1.0.1
Flask-Cors==5.0.0